backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
speechrecognition==3.10.0
pyttsx3==2.90
pdfplumber==0.10.3
scikit-learn==1.3.2
numpy==1.24.3
transformers==4.35.2
sentence-transformers==2.2.2
torch==2.1.1
httpx>=0.27.0
python-dotenv>=1.0.1
gtts>=2.3.2
openai>=1.14.0